// src/app/model/productModel.ts
export class productModel {
  itemName: string = "";
  image: string = "";
  info: string = "";
  cost: number = 0;
}


  // itemList = [
  //   {
  //     itemName : "Titan",
  //     cost : 3000,
  //     info : "One of the finest watches",
  //     image : "https://images.pexels.com/photos/190819/pexels-photo-190819.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1"
  //   },
  //   {
  //     itemName : "Tissot",
  //     cost : 10000,
  //     info : "One of the finest watches",
  //     image : "https://images.pexels.com/photos/2113994/pexels-photo-2113994.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1"
  //   },
  //   {
  //     itemName : "Sonata",
  //     cost : 1000,
  //           info : "One of the finest watches",
  //     image : "https://images.pexels.com/photos/277390/pexels-photo-277390.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1"
  //   },
  //   {
  //     itemName : "Rado",
  //     cost : 15000,
  //           info : "One of the finest watches",
  //     image : "https://images.pexels.com/photos/128206/pexels-photo-128206.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1"
  //   },
  //   {
  //     itemName : "Fastrack",
  //     cost : 1500,
  //           info : "One of the finest watches",
  //     image : "https://images.pexels.com/photos/280250/pexels-photo-280250.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1"
  //   },
  //   {
  //     itemName : "Rolex",
  //     cost : 15500,
  //           info : "One of the finest watches",
  //     image : "https://images.pexels.com/photos/1120275/pexels-photo-1120275.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1"
  //   },
  //   {
  //     itemName : "Casio",
  //     cost : 5500,
  //           info : "One of the finest watches",
  //     image : "https://images.pexels.com/photos/266666/pexels-photo-266666.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1"
  //   },
  // ]